Refactor: InteractionManager 10 modüle bölündü
InteractionManager dosyası (3528 satır) anlamlı bir şekilde 10 ayrı modüle bölündü:

📁 Yeni yapı:
- pointer/ (move, down, up - 456 satır)
- keyboard/ (keyboard-handler - 252 satır)
- ghost/ (ghost-updater - 205 satır)
- placement/ (component-placement - 555 satır)
- pipe/ (pipe-drawing - 400 satır)
- drag/ (drag-handler - 821 satır)
- rotation/ (rotation-handler - 183 satır)
- selection/ (selection-manager - 154 satır)
- helpers/ (finders - 610 satır)

🎯 Ana dosya: 3528 → 368 satır (89% azalma)

✅ Avantajlar:
- Her modül tek sorumlulukla 300-500 satır arası
- Daha kolay bakım ve debug
- Kod tekrarı azaltıldı
- İmport/export ile net bağımlılıklar

🔧 Orijinal dosya .backup olarak korundu
